chore: added types and refactored utils
