+++ apps/backend/src/types/bindings.d.ts
@@ -1,0 +1,9 @@
+export type Env = {
+  DATABASE_URL: string;
+  DATABASE_AUTH_TOKEN: string;
+  CHROMA_SERVER_URL: string;
+  CHROMA_USERNAME: string;
+  CHROMA_PASSWORD: string;
+  CF_EMBEDDING_API_KEY: string;
+  CF_EMBEDDING_MODEL: string;
+}
